DatabaseUpdater: Added updaters + removed replicators
Replicators where a dumb queue which sent messages through the
discovery protocol about database updates. If a burst of updates was
triggered, a bunch of UDP messages about the update were sent.

DatabaseUpdater is based on DatabaseAutocommit, which throttles
the number of messages. When a endpoint is not so busy (changing too
fast) it sends updates every 5s, when it's busy, every 15s; and it
has a maximum forced timeout of 75s.