Optimization: Make use of non-static inline.
The inline specifier was historically intended to be a code inlining
hint. However, it has a semantic effect functions marked inline can be
defined in every TU in which they are used (as long as the definitions
are identical): "inline" means "There will be multiple definitions of
this in different TUs; they're all the same, please treat them as one".