xapian-core: Remote Database directory selection support.
If the user requests to open a remote database server, it can
optionally require the remote backend to tell the server what database
it wants to open, if the remote server is configured as a server
without a default directory (xapian-tcpsrv remains to always have a
default directory at the moment).

The binary protocol was changed in the way that a new command can be
sent to the server (MSG_READACCESS) in case the server tells the
client no default directory is selected, the server then reacts to
this command by opening and selecting the requested database as the
active database.

* RemoteDatabase requires database index if server has none set.

* Remote server implements database selection through msg_writeaccess
    and msg_readaccess.